requirements.txt
google-generativeai>=0.3.0
pom-parser>=0.1.0
xmltodict>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2 